+++ src/i2c_handler.h
@@ -11,14 +11,14 @@ uint32_t i2c_handler_initialise(uint32_t baudrate);
 
 void i2c_handler_disable(void);
 
-int8_t i2c_handler_selectHwInstance(uint8_t device_num);
+int32_t i2c_handler_selectHwInstance(uint8_t device_num);
 
 uint32_t i2c_handler_set_baudrate(uint32_t baudrate);
 
 
-int8_t i2c_handler_write(uint8_t addr, uint8_t* buffer, uint32_t amount);
+int32_t i2c_handler_write(uint8_t addr, uint8_t* buffer, uint32_t amount);
 
-int8_t i2c_handler_read(uint8_t addr, uint8_t* buffer, uint32_t amount);
+int32_t i2c_handler_read(uint8_t addr, uint8_t* buffer, uint32_t amount);
